pyproject.toml
[tool.poetry]
package-mode = false
authors = ["Dmitrii Lavrukhin <lavrukhindmitry@gmail.com>"]
repository = "https://github.com/Eldies/image-processor"

[tool.poetry.dependencies]
python = "~3.14"
fastapi = ">=0.135"
uvicorn = ">=0.43.0"
pillow = ">=12.2.0"
rembg = {version = ">=2.0.74", extras = ["cpu"]}
python-multipart = ">=0.0.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
